Move export/import to mappings command
